+++ fractions_app/windows/congratulation_window.py
@@ -10,7 +10,7 @@ class CongratulationWindow(Toplevel):
         self.title("Прикинь ти")
         self.resizable(False, False)
 
-        width, height = master.winfo_width() - 250, master.winfo_height() - 250
+        width, height = master.winfo_width() - 250, master.winfo_height() - 180
         x, y = master.winfo_x() + 150, master.winfo_y() + 100
 
         self.geometry(f"{width}x{height}+{x}+{y}")
@@ -18,7 +18,7 @@ class CongratulationWindow(Toplevel):
         self.canvas = Canvas(self, width=self.img.width(), height=self.img.height())
         self.canvas.create_image(
             self.img.width(),
-            self.img.height() // 2,
+            self.img.height() // 2 + 20,
             image=self.img,
         )
         self.canvas.pack(fill="both", expand=True)