+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 1.0.1
+In this update:
+* Fixed the congratulation window image.
+* Fixed problem when result of operation is zero, now when result of operation is **0** user should pass fraction **0 <sup>0</sup>/<sub>0</sub>**.
+
 # Version 1.0.0
 In this update:
 * Changed the exercise window(now input fields are placed like fractions).